Osx and clang fixes

+++ common/writepcx.cpp
@@ -170,5 +170,5 @@ void Write_Pcx_ScanLine(int file_handle, int scansize, unsigned char* ptr)
         }
     }
 
-    Write_File(file_handle, pool, (int)file_ptr - (int)pool);
+    Write_File(file_handle, pool, file_ptr - pool);
 }
